Actually send the engine <-> app messages
The messages to the app were never sent because the queue was
never finished. Similarly, when the host set a value it didn't
call requestFlus so without processing, nothing worked.

So the parameters window works absent an audio loop with this changed!

Still not sure why on mac we don't get an audio callback. but that's
a separate issue